Move parsing text to Grammr.Text namespace.
