+++ specs/DotNetProjectFile.Analyzers.Specs/TestTools/Source.cs
@@ -1,4 +1,5 @@
 using DotNetProjectFile.Syntax;
+using Grammr.Text;
 using Microsoft.CodeAnalysis.Text;
 
 namespace Specs.TestTools;
