+++ src/DotNetProjectFile.Analyzers.Sdk/DotNetProjectFile.Analyzers.Sdk.props
@@ -41,7 +41,7 @@
 
   <!-- Ensure our analyzers are available -->
   <ItemGroup>
-    <PackageReference Include="DotNetProjectFile.Analyzers" Version="1.5.1" PrivateAssets="all" />
+    <PackageReference Include="DotNetProjectFile.Analyzers" Version="1.5.2" PrivateAssets="all" />
   </ItemGroup>
 
 </Project>
