+++ src/DotNetProjectFile.Analyzers/Syntax/SyntaxNode.cs
@@ -1,4 +1,5 @@
 using DotNetProjectFile.Collections;
+using Grammr.Text;
 using Microsoft.CodeAnalysis.Text;
 
 namespace DotNetProjectFile.Syntax;
