+++ src/DotNetProjectFile.Analyzers/Syntax/SyntaxTree.cs
@@ -1,3 +1,4 @@
+using Grammr.Text;
 using Microsoft.CodeAnalysis.Text;
 using System.IO;
 using System.Text;
